pyproject.toml
[tool.poetry]
name = "wiki-searcher"
version = "0.1.0"
description = ""
authors = ["Aneesh Anil <aneesh@geektrust.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.308"
openai = "^0.28.1"
google-search-results = "^2.4.2"
numexpr = "^2.8.7"
wikipedia = "^1.4.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
